Some CMake fixes/improvements (#1353)
* CMake: add FindONNXRuntime to allow for external ONNX to be found if available

Default installations (meaning non ALICE ones) of ONNXRuntime do not
provide a CMake Config file, but do provide a proper pkgconfig file.
So this cmake module first try to locate ONNXRuntime using the config
file (as before), but in case of failure it tries also the pkgconfig
file.

* CMake: add Findfjcontrib module file.

For some Fastjet installations, fjcontrib is actually a separate package
altogether. This module allows that. Note that it does not change the
existing behavior (where Fastjet and Fastjet contrib are installed, and
found, together), and the `find_package(fjcontrib)` will be a noop in
that case (provided the find_package(Fastjet) comes first).

* CMake: fix for o2physics_add_header_only_library

* CMake: move all dependencies finding in dependencies dir

And ensure the right modules are included before use.

Also add topdir to (default) include directory of targets to avoid
having to use `include_directories` (which is "old" cmake way of doing
things).

* CMake: (re)activate DataModel as a header only library